Rename IsomorphicApp in preparation for changing the case of the first character
This is a problem caused by MacOS being case insensitive
